+++ src/storage_manager.py
@@ -14,188 +14,74 @@ import re
 import catalog
 
 class StorageManager:
-
-
-
     def __init__(self, dbloc, pageSize, bufferSize) -> None:
         self.dbloc = dbloc
-        self.pageSize = pageSize
-        self.bufferSize = bufferSize
         self.cat = catalog.Catalog(self.dbloc, self.pageSize, self.bufferSize)
-        self.TYPE_LEN = 3
-        self.INT_BYTE_MAX_LEN = 7
-        self.INT_BYTE_TYPE = "big"
 
 
-    #creates an empty table
-    def create_table(self, table_name):
-        filepath = self.dbloc + "/" + table_name + ".bin"
-        #if there is not a table yet
-        if os.path.exists(filepath):
-            print(f"Table {table_name} already exists")
-            return 1
-        #create a new table
-        f = open(filepath, "wb+")
-        #write 0 as an integer to the table, there are currently 0 pages.
-        #f.write(int.to_bytes(0, self.INT_BYTE_MAX_LEN, self.INT_BYTE_TYPE))
-        #TODO fix out how to fix this error with bad returns. (return negative 1?) <-- Definitely not.
-        f.close()
-        return 0
+    """
+    • getting a record by primary key
+    • getting a page by table and page number
+    • getting all records for a given table number
+    • inserting a record into a given table
+    • deleting a record by primary key from a given table
+    • updating a record by primary key in a given table
+    """
+    #phase 1
+    #input records is a 2d array of records ex) [[12,"hello"],[5,"hey"]]
+    def insert_records(self, records):
+        # if there are no pages for this table:
+        # make a new file for the table
+        # add this entry to a new page
+        # insert the page into the table file
+        
+        # end
+        # Read each table page in order from the table file:
+        # iterate the records in page:
+        # if the record belongs before the current record:
+        # insert the record before it
+        # if the current page becomes overfull:
+        # split the page
+        # end
+        # If the record does not get inserted:
+        # insert it into the last page of the table file
+        # if page becomes overfull:
+        # split the page
+        # end
+        # splitting a page:
+        # make a new page
+        # remove half the items from the current page
+        # add the items to the new page
+        # insert the new page after the current page in the table
 
-    def get_record(self, primary_key):
-        return
-
-    def get_page(self, table_number, page_number):
         return
     
-    def bytes_to_int(self, byteRepresentation):
-        return int.from_bytes(byteRepresentation, self.INT_BYTE_TYPE)
+    #phase 1
+    def get_page(self, table_name, page_number):
+        #get from page buffer
+        return
 
     #phase 1
-    #returns a list of tuples ex: [(), (), ()][(), (), ()]
-    def get_records(self, table_name):
-        #get table attributes from catalog
-        attributes = self.cat.table_attributes(table_name)
-        records = []
-        #TODO if filepath does not exist, throw an error
-        filepath = self.dbloc + "/" + table_name + ".bin"
-        #open the filepath
-        with open(filepath, "rb") as f:
-            #read in the number of pages
-            num_pages = self.bytes_to_int(f.read(self.INT_BYTE_MAX_LEN))
-            #for each page, read in the number of records, then read each record
-            for i in range(num_pages):
-                num_records = self.bytes_to_int(f.read(self.INT_BYTE_MAX_LEN))
-                #for each record in the page, read each attribute
-                for j in range(num_records):
-                    record = []
-                    #for each attribute in the record
-                    for k in attributes:
-                        attribute_type = k["type"]
-                        #print(attribute_type)
-                        if attribute_type == 'integer':
-                            value = self.bytes_to_int(f.read(self.INT_BYTE_MAX_LEN))
-                        elif attribute_type == 'boolean':
-                            value = self.bytes_to_int(f.read(1))
-                            if value == 0:
-                                record.append("False")
-                            elif value == 1:
-                                record.append("True")
-                        elif attribute_type == 'double':
-                            binfloatone = f.read(self.INT_BYTE_MAX_LEN)
-                            floatone = int.from_bytes(binfloatone, self.INT_BYTE_TYPE)
-                            binfloattwo = f.read(self.INT_BYTE_MAX_LEN)
-                            floattwo = int.from_bytes(binfloattwo, self.INT_BYTE_TYPE)
-                            value = float(str(floatone) + "." + str(floattwo))
-                        #if fixed len char
-                            #char_len = TODO get the length of the char in bytes
-                                #this can be gotten from the attribute table from the catalog
-                            #value = f.read(char_len)
-                            #value = value.decode()
-                        #if varchar
-                            #knownInt = f.read(self.INT_BYTE_MAX_LEN)
-                            #knownInt = int.from_bytes(knownInt, self.INT_BYTE_TYPE)
-                            #print(knownInt)
-                            #knownString = f.read(knownInt)
-                            #print(knownString.decode())
-
-                        #attribute_len = get_len()
-                        #record.append(get_len())
-
-                        #add the record attribute
-                        record.append(value)
-                    records.append(record)
-            return records
-
-    #helper function to get the length to read
-    #attribute is the string of the current attribute being checked: ex)
-    def get_len(attribute_type, value):
+    def get_all_records(self, table_name):
         return
 
-    # phase 1
-    def insert_record(self, table_name, attributes, values):
-        #print(len(values))
-        if (len(attributes) != len(values)):
-            print("SM: Attribute size does not equal value size")
-            return 0
-
-        filepath = self.dbloc + "/" + table_name + ".bin"
-        #if there is not a table yet
-        if os.path.exists(filepath) is False:
-            print("SM: Table does not exist.")
-            #create a new table
-            #self.create_table(table_name)
-            
-        table_attributes = self.cat.table_attributes(table_name)
-        #traverse the file to find where this record belongs
-        #for each attribute in the table
-        for i in range(len(table_attributes)):
-            attribute = table_attributes[i]
-            #if the attribute is the primary key, store the primary index for later
-            if attribute["primary_key"]:
-                primary_index = i
-        #open the table file to read and write in binary
-        with open(filepath, "rb+") as f:
-            num_pages_pointer = f
-            #read in the number of pages
-            num_pages = self.bytes_to_int(f.read(self.INT_BYTE_MAX_LEN))
-            #if there are 0 pages, create a new page
-            #if num_pages == 0:
-            #increment num_pages to 1
-            num_pages += 1
-            #increment catalog page count
-            self.cat.update_page_count(table_name,1)
-            #print(int.to_bytes(num_pages, self.INT_BYTE_MAX_LEN, self.INT_BYTE_TYPE))
-            #write page count to the table
-            num_pages_pointer.write(int.to_bytes(num_pages, self.INT_BYTE_MAX_LEN, self.INT_BYTE_TYPE))
-                
-            #for each page, read in the number of records, then read each record
-            for i in range(num_pages):
-                num_records_pointer = f
-                num_records = self.bytes_to_int(f.read(self.INT_BYTE_MAX_LEN))
-                #if there are 0 records, write the value.
-                if num_records == 0:
-                    num_records+=1
-                    #write the number of records to the page
-                    num_records_pointer.write(int.to_bytes(1, self.INT_BYTE_MAX_LEN, self.INT_BYTE_TYPE))
-                    #print(values)
-                    #print(table_attributes)
-                    #write a record to this position
-                    toWrite = self.record_to_bytes(values,table_attributes)
-                    if toWrite == 1:
-                        print("Type error: we only have integer working")
-                        return 1
-                    f.write(toWrite)
-                    self.cat.update_record_count(table_name,1)
-                    #return successfully wrote record
-                    return 0
-                #for each record in the page, read each attribute
-                for j in range(num_records):
-                    record = []
-                    #for each attribute in the record
-            #once the end of the file is reached, append new record
-            
-
+    #phase 2
+    def alter_table(self, table_name):
         return
     
-    def record_to_bytes(self, values, attributes):
-        for j in range(len(values)):
-            #print(len(attributes))
-            for k in range(len(attributes)):
-                attribute_type = attributes[k]["type"]
-                #print(attribute_type)
-                if attribute_type == 'integer':
-                    #print(values[j][k])
-                    int_val = int.to_bytes(int(values[j][k]), self.INT_BYTE_MAX_LEN, self.INT_BYTE_TYPE)
-                else:
-                    return 1
-        return int_val
-    
-    def delete_record(self, primary_key):
+    #phase 2
+    def drop_table(self, table_name):
         return
 
-    def update_record(self, primary_key):
+    #phase 3
+    def delete_record(self, table_name, primary_key):
         return
+
+    #phase 3
+    def update_record(self, table_name, primary_key):
+        return
+    
+
 
 if __name__ == '__main__':
     SM = StorageManager("testDB", "1024", "64")
